+++ packages/ricos-types/web/src/themeTypes.ts
@@ -75,6 +75,10 @@ export interface CustomMentionStyle {
 export interface CustomFooterToolbarStyle {
   marginTop?: CSSProperties['marginTop'];
 }
+export interface CustomEditorStyle {
+  minHeight?: CSSProperties['minHeight'];
+  maxHeight?: CSSProperties['maxHeight'];
+}
 
 export interface RicosCustomStyles {
   h1?: CustomTextualStyle;
@@ -91,6 +95,7 @@ export interface RicosCustomStyles {
   codeBlock?: CustomCodeBlockStyle;
   mention?: CustomMentionStyle;
   footerToolbar?: CustomFooterToolbarStyle;
+  editor?: CustomEditorStyle;
 }
 
 export interface RicosSettingsStyles {